+++ demo/seminr-contained.R
@@ -9,12 +9,9 @@ mobi_pls <- estimate_pls(
     composite("Image", multi_items("IMAG", 1:5)),
     composite("Expectation", multi_items("CUEX", 1:3)),
     composite("Value", multi_items("PERV", 1:2)),
-    composite("Satisfaction", multi_items("CUSA", 1:3))
-  ),
-
-  interactions(
-    interaction_ortho("Image", "Expectation"),
-    interaction_ortho("Image", "Value")
+    composite("Satisfaction", multi_items("CUSA", 1:3)),
+    interaction_term("Image*Expectation", dimensions = c("Image","Expectation"), method = orthogonal),
+    interaction_term("Image*Value", dimensions = c("Image","Value"), method = orthogonal)
   ),
 
   relationships(