+++ demo/seminr-interaction.R
@@ -8,23 +8,21 @@ mobi_mm <- constructs(
   composite("Expectation",  multi_items("CUEX", 1:3)),
   composite("Value",        multi_items("PERV", 1:2)),
   composite("Satisfaction", multi_items("CUSA", 1:3)),
-  interaction(dimensions = c("Image","Expectation"), method = orthogonal)
+  interaction_term(name = "Image*Expectation", dimensions = c("Image","Expectation"), method = product_indicator),
+  interaction_term(name = "Image*Value", dimensions = c("Image","Expectation"), method = product_indicator)
 )
-
-
 
 # Structural model
 #  note: interactions should be the names of its main constructs joined by a '*' in between.
 mobi_sm <- relationships(
   paths(to = "Satisfaction",
         from = c("Image", "Expectation", "Value",
-                 "Image*Expectation"))
+                 "Image*Expectation","Image*Value"))
 )
 
 # Load data, assemble model, and estimate
 mobi_pls <- estimate_pls(data = mobi,
                          measurement_model = mobi_mm,
-                         interactions = NULL,
                          structural_model = mobi_sm)
 
 summary(mobi_pls)
@@ -41,13 +39,9 @@ mobi_mm <- constructs(
   composite("Image",        multi_items("IMAG", 1:5)),
   composite("Expectation",  multi_items("CUEX", 1:3)),
   composite("Value",        multi_items("PERV", 1:2)),
-  composite("Satisfaction", multi_items("CUSA", 1:3))
-)
-
-# interaction constructs must be created after the measurement model is defined
-mobi_xm <- interactions(
-  interaction_scaled("Image", "Expectation"),
-  interaction_scaled("Image", "Value")
+  composite("Satisfaction", multi_items("CUSA", 1:3)),
+  interaction_term(name = "Image*Expectation", dimensions = c("Image","Expectation"), method = product_indicator),
+  interaction_term(name = "Image*Value", dimensions = c("Image","Expectation"), method = product_indicator)
 )
 
 # structural model: note that name of the interactions construct should be
@@ -61,7 +55,6 @@ mobi_sm <- relationships(
 # Load data, assemble model, and estimate
 mobi_pls <- estimate_pls(data = mobi,
                          measurement_model = mobi_mm,
-                         interactions = mobi_xm,
                          structural_model = mobi_sm)
 
 summary(mobi_pls)